feat (DPLAN-17918): Move introduced code to a dedicated service
