+++ demosplan/DemosPlanCoreBundle/Logic/CustomField/CustomFieldFilterResolver.php
@@ -14,14 +14,18 @@ namespace demosplan\DemosPlanCoreBundle\Logic\CustomField;
 
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Segment;
 use demosplan\DemosPlanCoreBundle\Entity\Statement\Statement;
+use demosplan\DemosPlanCoreBundle\Logic\Statement\ElasticSearchService;
 use demosplan\DemosPlanCoreBundle\Utils\CustomField\Enum\CustomFieldSupportedEntity;
 use Doctrine\ORM\EntityManagerInterface;
+use Elastica\Query\AbstractQuery;
+use Illuminate\Support\Collection;
 use InvalidArgumentException;
 
 class CustomFieldFilterResolver
 {
     public function __construct(
         private readonly EntityManagerInterface $entityManager,
+        private readonly ElasticSearchService $elasticSearchService,
     ) {
     }
 
@@ -72,6 +76,49 @@ class CustomFieldFilterResolver
         }
 
         return array_column($query->getArrayResult(), 'id');
+    }
+
+    /**
+     * Strips customField_* entries from $userFilters, resolves matching statement
+     * IDs via Doctrine and returns the Elastica terms filter alongside the cleaned filters array.
+     *
+     * @param array<string, mixed> $userFilters
+     *
+     * @return array{0: AbstractQuery|null, 1: array<string, mixed>}
+     */
+    public function resolveCustomFieldFilter(string $procedureId, array $userFilters): array
+    {
+        $prefix = 'customField_';
+
+        /** @var Collection<string, mixed> $customFieldEntries */
+        /** @var Collection<string, mixed> $remainingFilters */
+        [$customFieldEntries, $remainingFilters] = collect($userFilters)->partition(
+            static fn (mixed $value, string $key): bool => str_starts_with($key, $prefix)
+        );
+
+        if ($customFieldEntries->isEmpty()) {
+            return [null, $userFilters];
+        }
+
+        $fieldFilters = $customFieldEntries
+            ->mapWithKeys(
+                static fn (mixed $value, string $key): array => [substr($key, strlen($prefix)) => $value]
+            )
+            ->toArray();
+
+        $matchingIds = $this->resolveMatchingIds(
+            CustomFieldSupportedEntity::statement,
+            $procedureId,
+            $fieldFilters
+        );
+
+        return [
+            $this->elasticSearchService->getElasticaTermsInstance(
+                'id',
+                [] !== $matchingIds ? $matchingIds : ['__no_match__']
+            ),
+            $remainingFilters->toArray(),
+        ];
     }
 
     /**
